requirements.txt
attrs==19.3.0
Automat==20.2.0
backcall==0.1.0
beautifulsoup4==4.8.2
betamax==0.8.1
bleach==3.1.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
decorator==4.4.1
defusedxml==0.6.0
demjson==2.2.4
entrypoints==0.3
Flask==1.1.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.3.0
incremental==17.5.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
js2xml==0.3.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
lxml==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.0.2
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.2
pickleshare==0.7.5
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
Protego==0.1.16
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.5.2
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
pyrsistent==0.15.7
python-dateutil==2.8.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
queuelib==1.5.0
requests==2.23.0
Scrapy==1.8.0
scrapyrt==0.11.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.13.0
slimit==0.8.1
soupsieve==1.9.5
SQLAlchemy==1.3.13
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
Twisted==19.10.0
urllib3==1.25.9
w3lib==1.21.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==0.6.0
zope.interface==4.7.1
